refactor(auth): bound submission preparation structure
