+++ backend/app/modules/artifacts/authorization.py
@@ -55,13 +55,15 @@ from app.modules.authorization.runtime import (
     GuideSourceReadResourceContext,
     PreSubmitCheckerInputResourceContext,
     PreSubmitCheckerInputPreparationContext,
-    SubmissionBundlePreparationResourceContext,
-    SubmissionBundlePreparationPreflightResourceContext,
     HumanAuthorizationContext,
     PreparedAuthorizationHandleInvalid,
     PreparedAuthorizationUnsupported,
     ServiceAuthorizationContext,
     AuthorizationContext,
+)
+from app.modules.authorization.submission_preparation import (
+    SubmissionBundlePreparationPreflightResourceContext,
+    SubmissionBundlePreparationResourceContext,
 )
