+++ backend/app/modules/authorization/runtime.py
@@ -19,6 +19,10 @@ from app.modules.authorization.domain.project_create import ProjectCreateResourc
 from app.modules.actors.service_identities import ServiceIdentity
 from app.modules.authorization.catalogue import ActionId, PermissionId
 from app.modules.authorization.schemas import AdminRole, AdminScope, ProjectRole
+from app.modules.authorization.submission_preparation import (
+    SubmissionBundlePreparationPreflightResourceContext,
+    SubmissionBundlePreparationResourceContext,
+)
 
 _STRICT_FROZEN = ConfigDict(extra="forbid", frozen=True, strict=True)
 PROJECT_DIAGNOSTIC_TARGET_KIND_BY_ACTION = {
@@ -1474,67 +1478,6 @@ class PreSubmitCheckerInputResourceContext(PreSubmitCheckerInputPreparationConte
     semantic_manifest_sha256: str = Field(pattern=r"^sha256:[0-9a-f]{64}$")
 
 
-class SubmissionBundlePreparationRequestContext(BaseModel):
-    """Server-owned contributor selectors known before reading request bytes."""
-
-    model_config = _STRICT_FROZEN
-
-    scope_project_id: UUID
-    actor_profile_id: UUID
-    identity_link_id: UUID
-    task_id: UUID
-    assignment_id: UUID
-    predecessor_submission_id: UUID | None
-
-
-class SubmissionBundlePreparationPreflightResourceContext(
-    SubmissionBundlePreparationRequestContext
-):
-    """Exact project-scoped selectors revalidated before request bytes are read."""
-
-    resource_type: Literal["submission_bundle_preparation_preflight"]
-    resource_id: UUID
-
-
-class SubmissionBundlePreparationResourceContext(SubmissionBundlePreparationRequestContext):
-    """Exact contributor, locked policy, evidence, and durable-intent facts."""
-
-    model_config = _STRICT_FROZEN
-
-    resource_type: Literal["submission_bundle_preparation"]
-    resource_id: UUID
-    predecessor_submission_version: int | None = Field(default=None, ge=1)
-    pre_submit_evidence_set_id: UUID
-    prepared_generation_id: UUID
-    guide_id: UUID
-    guide_version: str
-    source_snapshot_id: UUID
-    source_snapshot_sha256: str = Field(pattern=r"^sha256:[0-9a-f]{64}$")
-    effective_policy_id: UUID
-    effective_policy_sha256: str = Field(pattern=r"^sha256:[0-9a-f]{64}$")
-    pre_submit_policy_id: UUID
-    pre_submit_policy_sha256: str = Field(pattern=r"^sha256:[0-9a-f]{64}$")
-    effective_plan_sha256: str = Field(pattern=r"^sha256:[0-9a-f]{64}$")
-    semantic_manifest_id: UUID
-    semantic_manifest_sha256: str = Field(pattern=r"^sha256:[0-9a-f]{64}$")
-    archive_sha256: str = Field(pattern=r"^sha256:[0-9a-f]{64}$")
-    archive_byte_count: int = Field(ge=0)
-    media_type: Literal["application/zip"]
-    storage_scheme: Literal["local", "s3"]
-    operation_identity: str = Field(min_length=1)
-    replay_durable_intent_id: UUID | None
-
-    @model_validator(mode="after")
-    def require_exact_identity_and_predecessor(self):
-        if self.resource_id != self.prepared_generation_id:
-            raise ValueError("submission preparation resource must match generation")
-        if (self.predecessor_submission_id is None) != (
-            self.predecessor_submission_version is None
-        ):
-            raise ValueError("submission predecessor identity is incomplete")
-        return self
-
-
 AuthorizationResourceContext = (
     ActorSelfResourceContext
     | ProjectReadResourceContext
@@ -1733,7 +1676,7 @@ class AuthorizationDecision(BaseModel):
 
 
 class AuthorizationDenied(Exception):
-    """Fail-closed control flow carrying only one bounded decision."""
+    """Bounded denial control flow."""
 
     def __init__(self, decision: AuthorizationDecision) -> None:
         if decision.allowed or decision.denial_code is None:
@@ -1743,7 +1686,6 @@ class AuthorizationDenied(Exception):
 
     @property
     def public_code(self) -> str:
-        """Map internal catalogue outcomes to the stable public denial."""
         denial_code = self.decision.denial_code
         if denial_code is None:
             raise RuntimeError("authorization denial lost its denial code")
@@ -1756,4 +1698,4 @@ class AuthorizationDenied(Exception):
 
 
 class AuthorizationEvidenceUnavailable(RuntimeError):
-    """Authorization evidence could not be persisted safely."""
+    """Authorization evidence was unavailable."""
